Stylistic pass and deps updates
