+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "quadtree_rs"
 version = "0.1.3"
 authors = ["James Adam Buckland <james.adam.buckland@gmail.com>"]
-edition = "2018"
+edition = "2021"
 description = "Point/region Quadtree with support for overlapping regions."
 
 publish = true
@@ -26,9 +26,9 @@ license = "Apache-2.0"
 maintenance = { status = "actively-developed" }
 
 [dependencies]
-num = "0.2"
-derive_builder = "0.7"
-serde = { version = "1.0.152", features = ["derive"], optional=true}
+num = "0.4"
+derive_builder = "0.12"
+serde = { version = "1", features = ["derive"], optional = true }
 
 [features]
 serde = ["dep:serde"]
@@ -39,5 +39,9 @@ serde = ["dep:serde"]
 [dev-dependencies.cargo-husky]
 version = "1"
 default-features = false # Disable features which are enabled by default
-features = ["precommit-hook", "run-cargo-test", "run-cargo-clippy", "run-cargo-fmt"]
-
+features = [
+    "precommit-hook",
+    "run-cargo-test",
+    "run-cargo-clippy",
+    "run-cargo-fmt",
+]
