+++ src/entry.rs
@@ -15,13 +15,11 @@
 //! A view into a single entry in the Quadtree.
 // Influenced by https://doc.rust-lang.org/std/collections/hash_map/enum.Entry.html.
 
+use crate::{area::Area, point::Point};
+use num::PrimInt;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
-use {
-    crate::{area::Area, point::Point},
-    num::PrimInt,
-    std::default::Default,
-};
+use std::default::Default;
 
 /// A region/value association in the [`Quadtree`].
 ///
