+++ tests/query_tests.rs
@@ -16,10 +16,8 @@ mod util; // For unordered_elements_are.
 
 // For testing .query(), .modify().
 mod query_tests {
-    use {
-        crate::util::unordered_elements_are,
-        quadtree_rs::{area::AreaBuilder, Quadtree},
-    };
+    use crate::util::unordered_elements_are;
+    use quadtree_rs::{AreaBuilder, Quadtree};
 
     #[test]
     fn query_empty() {
