+++ tests/util.rs
@@ -1,8 +1,6 @@
-use {
-    num::{cast::FromPrimitive, PrimInt},
-    quadtree_rs::area::AreaBuilder,
-    std::{collections::HashSet, default::Default, fmt::Debug, hash::Hash, iter::FromIterator},
-};
+use num::{cast::FromPrimitive, PrimInt};
+use quadtree_rs::{AreaBuilder, Quadtree};
+use std::{collections::HashSet, default::Default, fmt::Debug, hash::Hash, iter::FromIterator};
 
 // Inspired by google/googletest's UnorderedElementsAre().
 // https://github.com/google/googletest/blob/master/googlemock/docs/CheatSheet.md#container-matchers
@@ -22,7 +20,7 @@ where
 }
 
 #[allow(dead_code)]
-pub fn print_quadtree<U, V>(qt: &quadtree_rs::Quadtree<U, V>)
+pub fn print_quadtree<U, V>(qt: &Quadtree<U, V>)
 where
     U: PrimInt + Default + FromPrimitive + Debug,
     V: Debug,
